test(api): share the alternate-identity auth fixtures via conftest
`as_environment_b` (a second environment in the same workspace) and the
member-vs-API-key auth switcher were defined inside test_builds.py, but
the environment boundary and the workspace-role admin gate need testing
wherever a write endpoint is added — not just there. Moved to conftest
unchanged; the switcher is renamed `role_auth_switcher` since it is no
longer specific to the concurrency-limit endpoints.
