+++ app/stardag-api/tests/test_builds.py
@@ -3022,62 +3022,6 @@ async def test_evicted_then_worker_completes_sticky_completed(client: AsyncClien
     assert holders["holders"] == []
 
 
-@pytest.fixture
-async def as_environment_b(async_engine):
-    """Context manager switching the app's auth override to a SECOND
-    environment in the same workspace (the tenancy boundary for
-    tasks/limits is the environment)."""
-    import contextlib
-    from uuid import UUID
-
-    from sqlalchemy.ext.asyncio import async_sessionmaker
-
-    from stardag_api.auth import SdkAuth, require_sdk_auth
-    from stardag_api.main import app
-    from stardag_api.models import Environment, User
-    from tests.conftest import DEFAULT_USER_ID, DEFAULT_WORKSPACE_ID
-
-    env_b_id = UUID("00000000-0000-0000-0000-00000000000b")
-    session_maker = async_sessionmaker(async_engine, expire_on_commit=False)
-    async with session_maker() as session:
-        session.add(
-            Environment(
-                id=env_b_id,
-                workspace_id=DEFAULT_WORKSPACE_ID,
-                name="Environment B",
-                slug="env-b",
-            )
-        )
-        await session.commit()
-
-    auth_b = SdkAuth(
-        environment=Environment(
-            id=env_b_id, workspace_id=DEFAULT_WORKSPACE_ID, name="Environment B"
-        ),
-        workspace_id=DEFAULT_WORKSPACE_ID,
-        user=User(
-            id=DEFAULT_USER_ID,
-            external_id="default-local-user",
-            email="default@localhost",
-            display_name="Default User",
-        ),
-    )
-
-    async def override_require_sdk_auth_b() -> SdkAuth:
-        return auth_b
-
-    @contextlib.contextmanager
-    def _switch():
-        previous = app.dependency_overrides[require_sdk_auth]
-        app.dependency_overrides[require_sdk_auth] = override_require_sdk_auth_b
-        try:
-            yield
-        finally:
-            app.dependency_overrides[require_sdk_auth] = previous
-
-    return _switch
-
-
 @pytest.mark.asyncio
 async def test_holders_and_evict_environment_isolation(
     client: AsyncClient, as_environment_b
@@ -3112,83 +3056,9 @@ async def test_holders_and_evict_environment_isolation(
 # ---------------------------------------------------------------------------
 
 
-@pytest.fixture
-async def limits_auth_switcher(async_engine):
-    """Context-manager factory switching the app's auth override between a
-    MEMBER-role user and an API-key (machine) credential in the default
-    environment. The default ``client`` auth (OWNER-role user) is restored
-    on exit."""
-    import contextlib
-    from uuid import UUID
-
-    from sqlalchemy.ext.asyncio import async_sessionmaker
-
-    from stardag_api.auth import SdkAuth, require_sdk_auth
-    from stardag_api.main import app
-    from stardag_api.models import Environment, User, WorkspaceMember
-    from stardag_api.models.enums import WorkspaceRole
-    from tests.conftest import DEFAULT_ENVIRONMENT_ID, DEFAULT_WORKSPACE_ID
-
-    member_user_id = UUID("00000000-0000-0000-0000-0000000000ae")
-    session_maker = async_sessionmaker(async_engine, expire_on_commit=False)
-    async with session_maker() as session:
-        session.add(
-            User(
-                id=member_user_id,
-                external_id="member-user",
-                email="member@localhost",
-                display_name="Member User",
-            )
-        )
-        session.add(
-            WorkspaceMember(
-                workspace_id=DEFAULT_WORKSPACE_ID,
-                user_id=member_user_id,
-                role=WorkspaceRole.MEMBER,
-            )
-        )
-        await session.commit()
-
-    environment = Environment(
-        id=DEFAULT_ENVIRONMENT_ID,
-        workspace_id=DEFAULT_WORKSPACE_ID,
-        name="Default Environment",
-    )
-    member_auth = SdkAuth(
-        environment=environment,
-        workspace_id=DEFAULT_WORKSPACE_ID,
-        user=User(
-            id=member_user_id,
-            external_id="member-user",
-            email="member@localhost",
-            display_name="Member User",
-        ),
-    )
-    # API-key auth context: no user attached (machine credential).
-    api_key_auth = SdkAuth(
-        environment=environment,
-        workspace_id=DEFAULT_WORKSPACE_ID,
-        user=None,
-    )
-
-    @contextlib.contextmanager
-    def _as(auth: SdkAuth):
-        async def _override() -> SdkAuth:
-            return auth
-
-        previous = app.dependency_overrides[require_sdk_auth]
-        app.dependency_overrides[require_sdk_auth] = _override
-        try:
-            yield
-        finally:
-            app.dependency_overrides[require_sdk_auth] = previous
-
-    return {"member": lambda: _as(member_auth), "api_key": lambda: _as(api_key_auth)}
-
-
 @pytest.mark.asyncio
 async def test_limit_writes_admin_gated_for_users(
-    client: AsyncClient, limits_auth_switcher
+    client: AsyncClient, role_auth_switcher
 ):
     """PUT/DELETE /concurrency-limits/{key} and evict require the workspace
     ADMIN role on the user (JWT) auth path; reads stay member-level."""
@@ -3200,7 +3070,7 @@ async def test_limit_writes_admin_gated_for_users(
         await client.put("/api/v1/concurrency-limits/gate", json={"max_concurrent": 2})
     ).status_code == 200
 
-    with limits_auth_switcher["member"]():
+    with role_auth_switcher["member"]():
         # Reads stay member-level.
         assert (await client.get("/api/v1/concurrency-limits")).status_code == 200
         holders = await client.get("/api/v1/concurrency-limits/gate/holders")
@@ -3225,7 +3095,7 @@ async def test_limit_writes_admin_gated_for_users(
     assert limits == [{"key": "gate", "max_concurrent": 2}]
 
     # API-key auth (machine credential): full access.
-    with limits_auth_switcher["api_key"]():
+    with role_auth_switcher["api_key"]():
         assert (
             await client.put(
                 "/api/v1/concurrency-limits/gate", json={"max_concurrent": 5}
